requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.1
Flask-Login>=0.6.3
Flask-WTF>=1.2.1
Flask-Migrate>=4.0.5
Flask-Mail>=0.9.1
Flask-Limiter>=3.5.0
gunicorn>=21.2.0
psycopg2-binary>=2.9.9;platform_system!="Darwin" or platform_machine!="arm64"
psycopg[binary]>=3.1.0;platform_system=="Darwin" and platform_machine=="arm64"
pandas>=2.1.3
openpyxl>=3.1.2
python-dotenv>=1.0.0
bcrypt>=4.1.1
email-validator>=2.1.0
mailjet-rest>=1.3.4
requests>=2.31.0
pytest>=7.4.0
pytest-flask>=1.3.0
